requirements.txt
torch>=2.0
arxiv==2.3.1
beautifulsoup4==4.14.3
einops==0.8.1
feedparser==6.0.12
matplotlib==3.8.4
numpy==1.26.4
openai==2.14.0
pandas==2.1.4
Pillow==10.4.0
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.2
tqdm==4.67.1
transformers==4.39.3

